@@ -6,7 +6,7 @@
 
     <groupId>dk.dtu.compute.se.pisd</groupId>
     <artifactId>roborally</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
+    <version>1.1.0-SNAPSHOT</version>
     <description>
         RoboRally demo with database access and loading
         some game boards from a files for the course
@@ -15,9 +15,10 @@
     </description>
 
     <properties>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.source>15</maven.compiler.source>
+        <maven.compiler.target>15</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <javafx.version>15.0.1</javafx.version>
     </properties>
 
     <dependencies>
@@ -58,6 +59,11 @@
             <version>5.4.2</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.openjfx</groupId>
+            <artifactId>javafx-controls</artifactId>
+            <version>${javafx.version}</version>
+        </dependency>
     </dependencies>
 
     <build>
@@ -83,7 +89,7 @@
                         <manifest>
                             <addClasspath>true</addClasspath>
                             <classpathPrefix>libs/</classpathPrefix>
-                            <mainClass>dk.dtu.compute.se.pisd.roborally.RoboRally</mainClass>
+                            <mainClass>dk.dtu.compute.se.pisd.roborally.StartRoboRally</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
@@ -116,6 +122,16 @@
                     <overview>src/main/java/dk/dtu/compute/se/pisd/overview.html</overview>
                 </configuration>
             </plugin>
+            <!-- With the additional class StartRoboRally, we do not need that plugin any more
+                 to start the application. It can be directly started in IntelliJ now.
+            <plugin>
+                <groupId>org.openjfx</groupId>
+                <artifactId>javafx-maven-plugin</artifactId>
+                <version>0.0.5</version>
+                <configuration>
+                    <mainClass>dk.dtu.compute.se.pisd.roborally.RoboRally</mainClass>
+                </configuration>
+            </plugin> -->
         </plugins>
     </build>
 
