@@ -128,13 +128,6 @@ public class Board extends Subject {
         }
     }
 
-    public void getNextPlayer() {
-        int number = getPlayerNumber(current);
-        Player next = getPlayer((number + 1) % getPlayersNumber());
-        setCurrentPlayer(next);
-        setCount(getCount() +1);
-    }
-
     public Phase getPhase() {
         return phase;
     }
@@ -190,52 +183,25 @@ public class Board extends Subject {
         int x = space.x;
         int y = space.y;
         switch (heading) {
-            case SOUTH:
-                y = (y + 1) % height;
-                break;
-            case WEST:
-                x = (x + width - 1) % width;
-                break;
-            case NORTH:
-                y = (y + height - 1) % height;
-                break;
-            case EAST:
-                x = (x + 1) % width;
-                break;
+            case SOUTH -> y = (y + 1) % height;
+            case WEST -> x = (x + width - 1) % width;
+            case NORTH -> y = (y + height - 1) % height;
+            case EAST -> x = (x + 1) % width;
         }
 
         return getSpace(x, y);
     }
 
     public String getStatusMessage() {
-        // This is actually a view aspect, but for making the first task easy for
+        // this is actually a view aspect, but for making assignment V1 easy for
         // the students, this method gives a string representation of the current
         // status of the game
 
-        // TODO Assignment V1: this string could eventually be refined
-        //      The status line should show more information based on
-        //      situation; for now, introduce a counter to the Board,
-        //      which is counted up every time a player makes a move; the
-        //      status line should show the current player and the number
-        //      of the current move!
-        return "Player = " + getCurrentPlayer().getName() + " Counter = " + getCount();
+        // XXX: V2 changed the status so that it shows the phase, the player and the step
+        return "Phase: " + getPhase().name() +
+                ", Player = " + getCurrentPlayer().getName() +
+                ", Step: " + getStep();
     }
 
-    // TODO Assignment V1: add a counter along with a getter and a setter, so the
-    //      state the board (game) contains the number of moves, which then can
-    //      be used to extend the status message including the number of
 
-    private int count;
-
-
-    public int getCount() {
-        return count;
-    }
-
-    public void setCount(int count) {
-        if (this.count != count) {
-            this.count = count;
-            notifyChange();
-        }
-    }
 }
